# openapi/opal-osa-integration.yaml
openapi: 3.0.0
info:
  title: OPAL Decision & Diagnostics API
  version: 1.0.0
  description: |
    Complete API specification for OPAL → OSA data flow, decision-making, and troubleshooting.

    This API maintains backward compatibility with existing OPAL integrations while providing
    enhanced decision layer capabilities with evidence-based recommendations, confidence scoring,
    and impact/effort analysis.

    **Key Features:**
    - Idempotent webhook processing for reliable data flow
    - Evidence-based decision making with confidence metrics
    - Comprehensive diagnostics and troubleshooting tools
    - Force sync and orchestration capabilities
    - Backward compatible with existing OPAL workflows
  contact:
    name: OPAL Development Team
    url: https://opal-2025.vercel.app
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://opal-2025.vercel.app
    description: Production server
paths:
...
  /api/orchestrations/replay:
    post:
      tags:
        - Orchestration
      summary: Replay failed or incomplete workflow executions
      description: |
        Allows replaying of failed or incomplete workflow executions for recovery
        and troubleshooting purposes. Maintains idempotency and event ordering.
      operationId: replayWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowReplayRequest'
      responses:
        '200':
          description: Replay initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowReplayResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Original workflow execution not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  schemas:
...
    OpalAgentType:
      type: string
      enum:
        - experiment_blueprinter
        - audience_suggester
        - content_review
        - roadmap_generator
        - integration_health
        - personalization_idea_generator
        - cmp_organizer
        - customer_journey
        - geo_audit
      description: Supported OPAL agent types
...
    WorkflowReplayRequest:
      type: object
      required:
        - execution_id
      properties:
        execution_id:
          type: string
          description: Original execution ID to replay
        from_offset:
          type: integer
          minimum: 0
          description: Starting offset for replay (0 = full replay)
          default: 0
        agent_filter:
          type: array
          items:
            $ref: '#/components/schemas/OpalAgentType'
          description: Specific agents to replay (empty = all agents)

    # Response Schemas
...
    WorkflowReplayResponse:
      allOf:
        - $ref: '#/components/schemas/StandardApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                replay_execution_id:
                  type: string
                  description: New execution ID for the replay
                original_execution_id:
                  type: string
                  description: Original execution ID that was replayed
                replay_status:
                  type: string
                  enum: [initiated, queued, running]
                events_to_replay:
                  type: integer
                  description: Number of events scheduled for replay
...
    StandardApiResponse:
      type: object
      required:
        - success
        - correlation_id
        - timestamp
      properties:
        success:
          type: boolean
          description: Whether the request was successful
          example: true
        data:
          description: Response data (varies by endpoint)
        message:
          type: string
          description: Optional human-readable message
        correlation_id:
          type: string
          description: Unique request identifier for tracing
          example: "api-1699876543210-def456"
        timestamp:
          type: string
          format: date-time
          description: Response generation timestamp (ISO 8601 format)
          example: "2025-11-12T15:45:00.000Z"
        duration_ms:
          type: integer
          minimum: 0
          description: Processing time in milliseconds
          example: 245
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
        - correlation_id
        - timestamp
      properties:
        success:
          type: boolean
          description: Always false for error responses
          example: false
        error:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error description
          example: "Invalid request body"
        correlation_id:
          type: string
          description: Unique request identifier for tracing
          example: "api-1699876543210-def456"
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp (ISO 8601 format)
          example: "2025-11-12T15:45:00.000Z"
        duration_ms:
          type: integer
          minimum: 0
          description: Processing time before error occurred
          example: 25
        data:
          description: Additional error details (only in development)
  securitySchemes:
    hmacSignature:
      type: apiKey
      in: header
      name: X-Signature-SHA256
      description: |
        HMAC-SHA256 signature for webhook authentication.

        Format: sha256=<hex_digest>

        The signature is computed using the raw request body and a shared secret key.
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "VALIDATION_ERROR"
            message: "Invalid request body"
            correlation_id: "api-1699876543210-def456"
            timestamp: "2025-11-12T15:45:00.000Z"
            duration_ms: 25
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "INTERNAL_SERVER_ERROR"
            message: "An unexpected error occurred"
            correlation_id: "api-1699876543210-def456"
            timestamp: "2025-11-12T15:45:00.000Z"
            duration_ms: 150
...